format and rename settings_helper.py
